made the main page experience more efficent by reducing down the shader material from two to one instance
