+++ app/globals.css
@@ -20,12 +20,6 @@ body {
   font-family: Arial, Helvetica, sans-serif;
 }
 
-@layer utilities {
-  .text-balance {
-    text-wrap: balance;
-  }
-}
-
 .icon-text-shadow {
   text-shadow: 2px 2px 4px rgba(0, 0, 0, 1);
 }
