+++ components/mainPageComponents/ReactIcon.tsx
@@ -2,44 +2,22 @@ import React, { useRef } from "react";
 import { useFrame } from "@react-three/fiber";
 import { useGLTF } from "@react-three/drei";
 import * as THREE from "three";
-import { useDarkMode } from "../providers/DarkModeProvider";
-import vertexShader from "./shaders/holographic/vertex.glsl";
-import fragmentShader from "./shaders/holographic/fragment.glsl";
-import random2D from "./shaders/includes/random2D.glsl";
 
-const ReactIcon = () => {
+type Props = {
+  holographicMaterial: THREE.ShaderMaterial;
+};
+
+const ReactIcon = ({ holographicMaterial }: Props) => {
   const icon = useRef<THREE.Group>(null);
-  const { darkMode } = useDarkMode();
   const { nodes } = useGLTF(
     "https://vazxmixjsiawhamofees.supabase.co/storage/v1/object/public/models/react-logo/model.gltf",
   );
 
-  const processedVertexShader = vertexShader.replace(
-    "#include ../includes/random2D",
-    random2D,
-  );
-
-  const lightColor = new THREE.Color(0xea580c);
-  const darkColor = new THREE.Color(0x22d3ee);
-  const color = darkMode ? darkColor : lightColor;
-
-  const holographicMaterial = new THREE.ShaderMaterial({
-    vertexShader: processedVertexShader,
-    fragmentShader: fragmentShader,
-    uniforms: {
-      uTime: new THREE.Uniform(0),
-      uColor: new THREE.Uniform(color),
-    },
-    side: THREE.DoubleSide,
-    blending: THREE.AdditiveBlending,
-  });
-
   useFrame(({ clock }) => {
     const elapsedTime = clock.getElapsedTime();
-    holographicMaterial.uniforms.uTime.value = elapsedTime * 0.75;
-
     const amplitude = Math.PI / 5;
     const frequency = 0.5;
+
     icon.current!.rotation.y = Math.sin(elapsedTime * frequency) * amplitude;
   });
