+++ package.json
@@ -15,6 +15,7 @@
     "schema": "./db/prisma/schema.prisma"
   },
   "dependencies": {
+    "@gsap/react": "^2.1.2",
     "@prisma/client": "^6.10.1",
     "@react-three/drei": "^9.122.0",
     "@react-three/fiber": "^8.18.0",
